ADD: Cross section plotting, reorganize plotting module
