+++ docs/source/API/XSectionDisplay.rst
@@ -1,0 +1,5 @@
+act.plotting.XSectionDisplay
+============================
+
+.. autoclass:: act.plotting.XSectionDisplay
+   :members:
