+++ docs/source/API/plotting.rst
@@ -8,6 +8,7 @@ a new Display object, please make it inherited from this class.
 :func:`act.plotting.TimeSeriesDisplay` handles the plotting of timeseries. 
 :func:`act.plotting.WindRoseDisplay` handles the plotting of wind rose plots.
 :func:`act.plotting.SkewTDisplay` handles the plotting of Skew-T diagrams.
+:func:`act.plotting.XSectionDisplay` handles the plotting of cross sections.
 
 .. toctree::
    :maxdepth: 2
@@ -16,4 +17,5 @@ a new Display object, please make it inherited from this class.
    TimeSeriesDisplay
    WindRoseDisplay
    SkewTDisplay
+   XSectionDisplay
