+++ requirements.txt
@@ -5,4 +5,6 @@ numpy
 scipy
 xarray
 astral
+cartopy
 dask
+distributed
